requirements.txt.txt
streamlit
tensorflow
opencv-python
numpy
fpdf
Pillow
